backend/requirements.txt
# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# Authentication
djangorestframework-simplejwt>=5.3.0

# Database
psycopg2-binary>=2.9.9

# File Processing & AI
Pillow>=10.1.0
pytesseract>=0.3.10
pdfplumber>=0.10.3
PyPDF2>=3.0.1
openai>=1.3.0

# Environment
python-dotenv>=1.0.0

# API Documentation
drf-yasg>=1.21.7

# Utilities
python-magic-bin>=0.4.14; sys_platform == 'win32'
